eliminating extra output checks if "Strict" template is provided
Till now, all line of code was checked if that would return anything and
if it's done on the last line of the chunk. With strict templates it's
not necessary with exponential performance boost in nested templates.
This way rapport has really minimal footprint and overhead! :)

Note to the universe: DO STRICT TEMPLATES!